refactor(base): Move JumpBounds implementation to internal module
